refactor(milestones): extract pure helpers out of useMilestone
Moves the two stateless module-level helpers (mapPollExhaustion and
sendOutputsAndDeliverables) into a sibling useMilestone.helpers.ts. This is a
behavior-preserving extraction that brings hooks/useMilestone.ts back under its
quality-gate baseline (1211 -> 1148 lines), clearing the 'grew while already
over the limit' regression. No logic changes; the touched unit suites and tsc
pass.
